code/requirements.txt
kafka-python==2.0.2
dash==2.14.1
plotly==5.15.0
pandas==2.0.3
redis==4.5.4
Flask==2.3.3
